@@ -43,66 +44,67 @@
   },
   "devDependencies": {
     "#": "link:./src/filters",
-    "@11ty/eleventy": "^1.0.0",
-    "@11ty/eleventy-img": "^2.0.0",
-    "@11ty/eleventy-plugin-rss": "^1.1.2",
-    "@11ty/eleventy-plugin-syntaxhighlight": "^4.0.0",
-    "@babel/core": "^7.17.9",
-    "@babel/preset-env": "^7.16.11",
+    "@11ty/eleventy": "^1.0.1",
+    "@11ty/eleventy-img": "^2.0.1",
+    "@11ty/eleventy-plugin-rss": "^1.2.0",
+    "@11ty/eleventy-plugin-syntaxhighlight": "^4.1.0",
+    "@babel/core": "^7.18.6",
+    "@babel/preset-env": "^7.18.6",
     "@rollup/plugin-babel": "^5.3.1",
-    "@rollup/plugin-commonjs": "^21.0.3",
-    "@rollup/plugin-node-resolve": "^13.1.3",
+    "@rollup/plugin-commonjs": "^22.0.1",
+    "@rollup/plugin-node-resolve": "^13.3.0",
     "@tommoor/remove-markdown": "^0.3.2",
-    "@typescript-eslint/parser": "^5.18.0",
-    "@yarnpkg/sdks": "^3.0.0-rc.1",
+    "@typescript-eslint/parser": "^5.30.5",
+    "@yarnpkg/sdks": "^3.0.0-rc.11",
     "accoutrement": "^4.0.1",
-    "autoprefixer": "^10.4.4",
-    "babel-jest": "^27.5.1",
+    "autoprefixer": "^10.4.7",
+    "babel-jest": "^28.1.2",
     "cascading-color-systems": "0.1.0-beta.15",
     "chalk": "^4.1.2",
     "chokidar-cli": "^3.0.0",
-    "core-js": "^3.21.1",
+    "core-js": "^3.23.4",
     "cssremedy": "0.1.0-beta.2",
     "date-fns": "^2.28.0",
-    "date-fns-tz": "^1.3.3",
+    "date-fns-tz": "^1.3.5",
     "dotenv": "^16.0.1",
     "doxray": "^0.10.1",
-    "eslint": "^8.13.0",
+    "eslint": "^8.19.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-import-resolver-node": "^0.3.6",
-    "eslint-import-resolver-typescript": "^2.7.1",
+    "eslint-import-resolver-typescript": "^3.2.5",
     "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-jest": "^26.1.4",
-    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-jest": "^26.5.3",
+    "eslint-plugin-prettier": "^4.2.1",
     "eslint-plugin-simple-import-sort": "^7.0.0",
-    "fs-extra": "^10.0.1",
-    "jest": "^27.5.1",
+    "fs-extra": "^10.1.0",
+    "jest": "^28.1.2",
+    "jest-environment-jsdom": "^28.1.2",
     "js-yaml": "^4.1.0",
     "lodash": "^4.17.21",
-    "markdown-it": "^12.3.2",
-    "markdown-it-anchor": "^8.6.2",
+    "markdown-it": "^13.0.1",
+    "markdown-it-anchor": "^8.6.4",
     "markdown-it-footnote": "^3.0.3",
     "markdown-it-mark": "^3.0.1",
     "mockdate": "^3.0.5",
     "netlify-plugin-11ty": "^1.1.0",
     "node-fetch": "^2.6.7",
     "npm-run-all": "^4.1.5",
-    "postcss": "^8.4.12",
-    "prettier": "^2.6.2",
+    "postcss": "^8.4.14",
+    "prettier": "^2.7.1",
     "rimraf": "^3.0.2",
-    "rollup": "^2.70.1",
+    "rollup": "^2.76.0",
     "rollup-plugin-terser": "^7.0.2",
-    "sass": "^1.50.0",
-    "sassdoc": "^2.7.3",
+    "sass": "^1.53.0",
+    "sassdoc": "^2.7.4",
     "sassdoc-theme-herman": "^4.0.2",
     "slugify": "^1.6.5",
     "striptags": "^3.2.0",
-    "stylelint": "^14.6.1",
+    "stylelint": "^14.9.1",
     "stylelint-config-prettier": "^9.0.3",
-    "stylelint-config-standard-scss": "^3.0.0",
+    "stylelint-config-standard-scss": "^4.0.0",
     "stylelint-prettier": "^2.0.0",
     "truncate-html": "^1.0.4",
-    "typescript": "^4.6.3",
+    "typescript": "^4.7.4",
     "typogr": "^0.6.8"
   },
   "packageManager": "yarn@3.2.1"
