@@ -1,6 +1,6 @@
 import ccs from 'cascading-color-systems';
 
-import { clickToCopy } from './clickToCopy';
+import { clickToCopy } from './clickToCopy.cjs';
 
 ccs();
 clickToCopy();
